Fix dropped netlink messages when too many for the buffer
Because the buffer size is a fixed 4kb, if a netlink message is
truncated, only some parts of it would end up in the buffer, leading to
the message being dropped. Instead, peek at the message size ahead via
MSG_PEEK|MSG_TRUNC and allocate a buffer to fit it dynamically. (AFAIK
netlink messages can be >4kb anyway, so this is a bit more generally
useful.)